% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_surv_R6.R
\name{outSurv}
\alias{outSurv}
\title{The R6 object of survival analysis}
\description{
The output of the survival analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-toDf}{\code{outSurv$toDf()}}
\item \href{#method-kmplot}{\code{outSurv$kmplot()}}
\item \href{#method-cidx_calc}{\code{outSurv$cidx_calc()}}
\item \href{#method-clone}{\code{outSurv$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toDf"></a>}}
\if{latex}{\out{\hypertarget{method-toDf}{}}}
\subsection{Method \code{toDf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{outSurv$toDf()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-kmplot"></a>}}
\if{latex}{\out{\hypertarget{method-kmplot}{}}}
\subsection{Method \code{kmplot()}}{
the method to plot km curve
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{outSurv$kmplot(outName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{outName}}{The output pptx name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{

}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cidx_calc"></a>}}
\if{latex}{\out{\hypertarget{method-cidx_calc}{}}}
\subsection{Method \code{cidx_calc()}}{
The method to calculate the c-index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{outSurv$cidx_calc()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{outSurv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
