% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_surv_R6_nomo.R
\name{outSurv_nomo}
\alias{outSurv_nomo}
\title{The output of predict of survNomogram}
\description{
The output of predict of survNomogram

The output of predict of survNomogram
}
\section{Super class}{
\code{\link[doSurvival:outSurv]{doSurvival::outSurv}} -> \code{outSurv_nomo}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{outSurv_nomo$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="doSurvival" data-topic="outSurv" data-id="cidx_calc">}\href{../../doSurvival/html/outSurv.html#method-cidx_calc}{\code{doSurvival::outSurv$cidx_calc()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="doSurvival" data-topic="outSurv" data-id="kmplot">}\href{../../doSurvival/html/outSurv.html#method-kmplot}{\code{doSurvival::outSurv$kmplot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="doSurvival" data-topic="outSurv" data-id="toDf">}\href{../../doSurvival/html/outSurv.html#method-toDf}{\code{doSurvival::outSurv$toDf()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{outSurv_nomo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
