% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_surv_R6_nomo.R
\name{survNomogram}
\alias{survNomogram}
\title{The survival nomogram R6 class}
\description{
create a R6 class for survival nomogram analysis
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{survNomogram$new()}}
\item \href{#method-run}{\code{survNomogram$run()}}
\item \href{#method-figure}{\code{survNomogram$figure()}}
\item \href{#method-predict}{\code{survNomogram$predict()}}
\item \href{#method-clone}{\code{survNomogram$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survNomogram$new(
  time,
  event,
  TimeInterval = c(12, 24, 36),
  funLabel = c("1-Years Survival Probability", "2-Years Survival Probability",
    "3-Years Survival Probability")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survNomogram$run(dt, restep = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-figure"></a>}}
\if{latex}{\out{\hypertarget{method-figure}{}}}
\subsection{Method \code{figure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survNomogram$figure(dt, outName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survNomogram$predict(dt)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survNomogram$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
