% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_surv_R6.R
\name{survRadiomics}
\alias{survRadiomics}
\title{R6 Class Representing a survival radiomics model}
\description{
radiomics-based survival model
}
\details{
Nothing
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{survRadiomics$new()}}
\item \href{#method-figure}{\code{survRadiomics$figure()}}
\item \href{#method-predict}{\code{survRadiomics$predict()}}
\item \href{#method-run}{\code{survRadiomics$run()}}
\item \href{#method-clone}{\code{survRadiomics$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survRadiomics$new(time, event, uni_p_thresh = 0.05, cor_p_thresh = 0.9)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-figure"></a>}}
\if{latex}{\out{\hypertarget{method-figure}{}}}
\subsection{Method \code{figure()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survRadiomics$figure(outName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survRadiomics$predict(dt.radiomics, dt.clinics)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survRadiomics$run(dt.radiomics, dt.clinics)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{survRadiomics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
